requirements.txt
playwright>=1.55.0
streamlit>=1.28.0
langchain-gradient
python-dotenv
